fix: remove enum, rename internal pkg to walker
