+++ .github/scripts/cmd/yacar_ci.go
@@ -6,7 +6,7 @@ import (
 
 	"github.com/coinhall/yacar/cmd/sorter"
 	"github.com/coinhall/yacar/cmd/validator"
-	"github.com/coinhall/yacar/internal/reader"
+	"github.com/coinhall/yacar/internal/walker"
 )
 
 func main() {
@@ -15,7 +15,7 @@ func main() {
 		log.Panicln("ROOT_DIR env var not set")
 	}
 
-	yacarFilePaths := reader.GetLocalYacarFiles(projRoot)
+	yacarFilePaths := walker.GetLocalYacarFiles(projRoot)
 
 	validator.Start(yacarFilePaths)
 	sorter.Start(yacarFilePaths)
